Fix text() error when editable field contains '\'
If you called text() on an editable field where the user has entered a
backslash but hasn’t completed a LaTeX template, the backslash
command’s text() method would throw an error when trying to call text()
on the MathBlock’s undefined ends[L]. Subsequent keystrokes would then
occur outside the template.

This commit ensures that the text() method works when an editable field
contains a backslash that hasn’t been converted into a LaTeX template.
If the editable field contains ‘\’ or part of a LaTeX command like
‘\sqr’, text() will return what the user has typed literally (‘\’ and
‘\sqr’, respectively).